# api.yaml
openapi: '3.0.3'
info:
  title: 'Test task'
  version: '0.0.1'
paths:
  /device:
    post:
      description: Adds a new device
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreationData'
      responses:
        '200':
          description: Returns device
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                  payload:
                    $ref: '#/components/schemas/AddDeviceResponse'
  /device/{serialNumber}:
    get:
      description: Get device data
      responses:
        default:
          description: Device data
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                  payload:
                    $ref: '#/components/schemas/Device'
  /device/:
    get:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceQuery'

      responses:
        default:
          description: Returns all devices that fit query.
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                  devices:
                    type: array
                    items:
                      $ref: '#/components/schemas/Device'
                  totalDevicesSize:
                    type: integer
  /event:
    post:
      description: Adds a new event for a specified device
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddEventInfo'
      responses:
        default:
          description: Returns api response
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string

  /event/device/{deviceId}:
    get:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatePageLlimit'
      responses:
        default:
          description: Returns all events that fit query
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                  totalEventsSize:
                    type: integer
  /event/grouped-by-device-type/:
    get:
      requestBody:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatePageLlimit'
      responses:
        default:
          description: Returns all events that fit query grouped by device type
          content:
                application/json:
                  schema:
                    properties:
                      message:
                        type: string
                      events:
                        type: object
                        additionalProperties:
                          items:
                            $ref: '#/components/schemas/Event'
                      totalEventsSize:
                        type: integer

components:
  schemas:
    AddDeviceResponse:
      type: object
      properties:
        secretKey:
          type: string
    DeviceCreationData:
      properties:
        name:
          type: string
        serial:
          type: number
          format: int64
        deviceType:
          type: string
    Device:
      properties:
        name:
          type: string
        serial:
          type: number
          format: int64
        deviceType:
          type: string
        isActive:
          type: boolean
        creationDate:
          type: string
          format: date-time
        firstActivityDate:
          type: string
          format: date-time
        lastActivityDate:
          type: string
          format: date-time
    AddEventInfo:
      properties:
        secretKey:
          type: string
        eventType:
          type: string
        deviceSerial:
          type: integer
          format: int32
        payload:
          type: object
    DeviceQuery:
      properties:
        type:
          type: string
        datePageLimit:
          $ref: '#/components/schemas/DatePageLlimit'
    DatePageLlimit:
      properties:
        latestCreationDate:
          type: string
          format: date-time
        earliestCreationDate:
          type: string
          format: date-time
        page:
          type: integer
    Event:
      properties:
        id:
          type: integer
        deviceId:
          type: integer
        type:
          type: string
        payload:
          type: object
        creationDate:
          type: string
          format: date-time
    ApiResponse:
      properties:
        message:
          type: string